requirements.txt
antiorm==1.2.1
customtkinter==5.0.3
darkdetect==0.7.1
db==0.1.1
db-sqlite3==0.0.1
future==0.18.2
Pillow==9.3.0
tkintertable==1.3.3
